Got data from api, need to append data to html and database
