+++ configuration/config.json
@@ -1,0 +1,21 @@
+  {
+  "development": {
+      "username": "root",
+      "password": "WARNING",
+      "database": "soccer_db",
+      "host": "127.0.0.1",
+      "dialect": "mysql"
+  },
+  "test": {
+    "username": "root",
+    "password": null,
+    "database": "testdb",
+    "host": "localhost",
+    "dialect": "mysql",
+    "logging": false
+  },
+  "production": {
+      "use_env_variable": "JAWSDB_URL",
+      "dialect": "mysql"
+  }
+}
