--- configuration/connection.js
@@ -1,22 +1,0 @@
-// Set up MySQL connection.
-var mysql = require("mysql");
-
-var connection = mysql.createConnection({
-  port: 3306,
-  host: process.env.DB_HOST,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASS,
-  database: "soccer_db"
-});
-
-// Make connection.
-connection.connect(function(err) {
-  if (err) {
-    console.error("error connecting: " + err.stack);
-    return;
-  }
-  console.log("connected as id " + connection.threadId);
-});
-
-// Export connection for our ORM to use.
-module.exports = connection;
