+++ views/layouts/main.handlebars
@@ -4,8 +4,8 @@
 <head>
   <meta charset="utf-8">
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta.3/css/bootstrap.min.css">
-  <link rel="stylesheet" href="/styles/styles.css" media="screen" title="no title" charset="utf-8">
-  <title>Home Page</title>
+  <link rel="stylesheet" href="/styles/styles.css" media="screen" title="Soccer Hooligans 365" charset="utf-8">
+  <title>Soccer Hooligans 365</title>
 </head>
 
 <body>
